Expanded the BEmu implementation to include more of the BB API.  - Much of the new implementations throw NotImplementedExceptions,  -  but more code written for the BB API will compile now than with previous BEmu version.
C#: Put in Constant, ConstantsList, Constraint, ConstraintsList, EventQueue, Operation, SchemaElementDefinition, SchemaTypeDefinition, and Exceptions.
 - Most of these throw NotImplementedExceptions.
 - Expanded CorrelationId, it can hold an object.
 - Expanded Datetime, Element, Name, Request, Schema, Subscription
 - Event.cs added GetEnumerator()
 - Expanded Session, can unsubscribe in more ways now.
 - Each element now has a TypeDefinition.

C++: Put in most of the new C# pieces.

Java: I need to port the C# changes to Java

Former-commit-id: 342fbd6f3a634fb111ca2b854cbaac2d9a8f5499

+++ BEmu_cpp/src/MarketDataRequest/MessageMarketSessionOpened.cpp
@@ -13,7 +13,7 @@ namespace BEmu
 {
 	namespace MarketDataRequest
 	{
-		MessageMarketSessionOpened::MessageMarketSessionOpened() : MessagePtr(Name("SessionStarted"), CorrelationId(0))
+		MessageMarketSessionOpened::MessageMarketSessionOpened() : MessagePtr(Name("SessionStarted"), CorrelationId((long)0))
 		{
 		}
